refactor(ui):重构ui，内部模块化，外部统一接口
