+++ src/stores/ui/sidebar/view.ts
@@ -1,0 +1,11 @@
+export const createViewModule = (storeContext: any) => {
+  const { currentView } = storeContext;
+
+  function isActiveView(viewId: string): boolean {
+    return currentView.value === viewId;
+  }
+
+  return {
+    isActiveView,
+  };
+};
